requirements.txt
langchain==0.2.17
langchain-core==0.2.43
langchain-community==0.2.19
langchain-openai==0.1.25
openai==1.109.1
python-dotenv==1.0.1
tiktoken==0.12.0
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
pypdf==4.2.0
beautifulsoup4==4.12.3
lxml==5.3.0
requests==2.32.5
gradio==5.49.1
chromadb==1.2.0
langfuse==2.40.0
rank-bm25==0.2.2
